requirements.txt
streamlit
pandas
google-cloud-bigquery
pyspark
yfinance
altair